requirements.txt
aiohttp
python-dotenv
typing-extensions
pyyaml
types-PyYAML
pytest
pytest-asyncio
pytest-cov
mypy
aiofiles
types-aiofiles
attrs
asyncpg
asyncpg-stubs
